ShowExceptions preferrs to send "text/plain"
Plain text is viewable anywhere and it should never be acceptable to send
HTML to a client that will not render it properly.

The real change here is the behaviour when responding to HTTP_ACCEPT "*/*"
which is typically sent by no graphical clients such as cURL.

Also includes DRYing up of tests and removing test loopping which is difficult
to debug.